Replaced write_log with parse_log
write_log only wrote the final file to disk after all the changes were made in another function.

It made no sense to have a separate function which just writes the files, so instead I have consolidated it in the parse_log function which does the actual conversion from bip to energymech format.

Right now all it does is to delete the date timestamp from each line in the log file.